Changed some members in the property serializer to be observer pointers since it is not its responsibility to manage the memory and furthermore, if not refptr had held it before the serializer got to it, the serializer would end up deleting it resulting in a crash. Cleaned up code style to reflect the style guide.

+++ delta3d/src/dtDirector/directortypefactory.cpp
@@ -137,7 +137,7 @@ namespace dtDirector
             else
             {
                mCurrentXMLParser++;
-               if (mCurrentXMLParser >= mXMLParserList.size())
+               if (mCurrentXMLParser >= (int)mXMLParserList.size())
                {
                   mXMLParserList.push_back(new DirectorParser());
                }
